change file name "StandardHeaderExtensions.cs" to "HeaderExtensions.cs"
